server/requirements.txt
alembic==1.7.5
aniso8601==7.0.0
autopep8==1.6.0
cffi==1.15.0
click==8.0.3
cryptography==35.0.0
falcon==3.0.1
graphene-federation==0.1.0
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.0.1
promise==2.3
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.27
SQLAlchemy-serializer==1.4.1
toml==0.10.2
zipp==3.6.0
